+++ meal-planner/src/index.js
@@ -1,12 +1,13 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 import App from './App'
-/* import GlobalStyle from './GlobalStyle'
+/* import GlobalStyle from './styles/GlobalStyle'
  */ import reportWebVitals from './reportWebVitals'
 
 ReactDOM.render(
     <React.StrictMode>
-        {/*   <GlobalStyle /> */}
+        {/*         <GlobalStyle />
+         */}{' '}
         <App />
     </React.StrictMode>,
     document.getElementById('root')
