+++ meal-planner/src/styles/StyleWrapper.js
@@ -1,0 +1,11 @@
+import React from 'react'
+import GlobalStyle from './GlobalStyle'
+
+export default function StyleWrapper({ children }) {
+    return (
+        <div>
+            <GlobalStyle />
+            {children}
+        </div>
+    )
+}
