+++ meal-planner/styleguide.config.js
@@ -1,7 +1,3 @@
 module.exports = {
     components: 'src/components/**/*.js',
-    ignore: ['**/App.js'],
-    defaultExample: true,
-    exampleMode: 'expand',
-    usageMode: 'expand',
 }
